Update SQL_DDL_Func_Rasa_Zozaite_HW.sql
Comment: CREATE VIEW is not enough to make your code rerunnable. The view needs to be located in the schema. Optional (just to think about): the logic may be simplified.
Answer: added DROP VIEW IF EXISTS and CREATE OR REPLACE VIEW. Added schema name „public“ to the code  public.sales_revenue_by_category_qtr.  Tried to simplify the code. 

Comment: What'll change if you comment this condition out?
Answer: Not much. Removed from the code.

Comment: I'd add it everywhere since I got errors due to the conflicts with other students' functions. Add IF EXISTS to make it fully reproducible.
Answer: added DROP FUNCTION IF EXISTS to this code and in other tasks.

Comment: It'd be better to add the same check selects to other objects in the task
Answer:  added error checks (country’s name is incorrect, missing) in 3 task;  added error checks (input is null value, title is missing, when case is different) in 4 task; added check (when language is not in the database, inserting with existing language in database).

Comment: What if country name is incorrect or not present in the database? Will the user understand the problem?
Answer: Very good point. For checking unknown or null input array, changed descriptions to RAISE EXCEPTION 'You must provide at least one country.' . Also added RAISE EXCEPTION 'The following countries do not exist in the database: %' in case when country name is wrong or not in the database.

Comment : What if the name is written in lower or upper case only?
Answer: changed to WHERE TRIM(LOWER(name)) = 'klingon'  here and in other places.

Comment: Probably more thorough exception handling is needed. I have a lot of function variants, and I got an error SQL Error [23505]: ERROR: duplicate key value violates unique constraint "film_pkey"
Detail: Key (film_id)=(1013) already exists.
Where: SQL statement "INSERT INTO film (
title,
<...>
Answer: I think film with id 1013 was already in database because I added it when checking how function works but did not deleted later. Added  exception for handling duplicate film values.

Comment: When I rerun this select, I get an error:
SQL Error [P0001]: ERROR: Movie "Vacations in the moon" already exists in film table.
Where: PL/pgSQL function new_movie5(text,text,year,integer,numeric,numeric,timestamp with time zone) line 7 at RAISE
Answer: It is because that movie is already in the film table. I added it when I tested the query but didn’t delete it afterward. However, after thinking more about the logic, I decided that I need to add at least one more exception, because there can be multiple films with the same name. To prevent inserting duplicate movies with the same title, I chose to use the year as an additional filter. This way, a movie with the same title can be added, but two movies with the same title and the same release year cannot be inserted.
WHERE title = p_title    AND release_year = p_release_year


Comment: Why exception and not notice if we have a default value?
Answer: rewrote this part of code and changed to RAISE NOTICE so that in case of null value, it was set default to Klingon language.
IF lang_id IS NULL THEN
	RAISE NOTICE 'Language "%" not found, default language set to Klingon.', p_language;
	SELECT language_id INTO lang_id
    FROM language
    WHERE TRIM(LOWER(name)) = 'klingon';
END IF;